[Improve] MyTagとUserTagで共通している部分をwrapperコンポーネントに切り出し
