+++ app/javascript/components/shared/TheTagWrapper.vue
@@ -1,0 +1,138 @@
+<template>
+  <div>
+    <!-- タブ -->
+    <the-tab :registered-tags="registeredTags" />
+    <!-- カレンダー -->
+    <the-calendar :tweet-dates="tweetDates" @input-date="fetchDateTweets" />
+    <v-container class="main-content d-flex flex-row-reverse pt-0" row>
+      <!-- ハッシュタグの情報 -->
+      <v-col class="hidden-sm-and-down" cols="12" md="4">
+        <v-card flat>
+          <tag-status
+            ref="tagStatus"
+            :registered-tag="registeredTag"
+            @push-delete="$emit('push-delete')"
+            @push-update="$emit('push-update', registeredTag)"
+            @push-cancel="fetchTagData"
+          />
+        </v-card>
+      </v-col>
+      <!-- ツイート -->
+      <v-col cols="12" md="8" class="pt-0">
+        <tweets-view :tweets="tweets" :user="user" />
+      </v-col>
+    </v-container>
+    <!-- ページネーション -->
+    <div class="text-center">
+      <v-pagination
+        v-model="page.currentPage"
+        :length="page.totalPages"
+        :total-visible="7"
+        @input="$changePaginationPage"
+      />
+    </div>
+  </div>
+</template>
+
+<script>
+import theCalendar from "../TheCalendar"
+import tagStatus from "../TagStatus"
+import theTab from "../TagsTab"
+import tweetsView from "../TagsTweets"
+
+export default {
+  components: {
+    theCalendar,
+    tagStatus,
+    theTab,
+    tweetsView
+  },
+  props: {
+    user: {
+      type: Object,
+      default: () => {},
+      required: true
+    },
+    registeredTags: {
+      type: Array,
+      default: () => [],
+      required: true
+    }
+  },
+  data() {
+    return {
+      page: {
+        currentPage: 1,
+        totalPages: 1,
+        requestUrl: ""
+      },
+      registeredTag: {
+        id: "",
+        tweetedDayCount: "",
+        privacy: "",
+        remindDay: "",
+        firstTweetedAt: "",
+        lastTweetedAt: "",
+        tag: {
+          name: ""
+        }
+      },
+      tweets: [],
+      tweetDates: []
+    }
+  },
+  computed: {
+    registeredTagUrl() {
+      const { tagId } = this.$route.params
+      return `/api/v1/registered_tags/${tagId}`
+    }
+  },
+  watch: {
+    $route() {
+      this.firstRead()
+    }
+  },
+  mounted() {
+    this.firstRead()
+  },
+  methods: {
+    async firstRead() {
+      this.fetchTweetDates()
+      this.fetchTweetsData()
+      await this.fetchTagData()
+      document.title = `#${this.registeredTag.tag.name} | Hashlog`
+    },
+    async fetchTagData() {
+      const registeredTagRes = await this.$axios.get(this.registeredTagUrl)
+      const { registeredTag } = registeredTagRes.data
+      this.registeredTag = registeredTag
+    },
+    async fetchTweetsData() {
+      const tweetsRes = await this.$axios.get(
+        `${this.registeredTagUrl}/tweets?page=1`
+      )
+      this.$setPaginationData(tweetsRes)
+      const { tweets } = tweetsRes.data
+      this.tweets = tweets
+    },
+    // カレンダー用全てのツイート
+    async fetchTweetDates() {
+      this.tweetDates = []
+      const { tagId } = this.$route.params
+      const tweetsRes = await this.$axios.get(`${this.registeredTagUrl}/tweets`)
+      const { tweets } = tweetsRes.data
+      tweets.forEach(t => this.tweetDates.push(t.tweetedAt.substr(0, 10)))
+      this.date = ""
+    },
+    // カレンダーの日付の変更
+    async fetchDateTweets(date) {
+      const tweetsRes = await this.$axios.get(
+        `${this.registeredTagUrl}/tweets?page=1&date=${date}`
+      )
+      this.$setPaginationData(tweetsRes)
+      const { tweets } = tweetsRes.data
+      this.tweets = tweets
+    }
+  }
+}
+</script>
